requirements.txt
fastApi
uvicorn
pydantic
pytest
psycopg2-binary
python-dotenv
pylint
PyYAML
ulid-py
pytest-cov
requests
httpx